+++ RELEASES.md
@@ -2,6 +2,7 @@ Version 1.0.1
 =============
 
 * Add `Integer::secure_pow_mod` methods.
+* Add `Integer::div_rem_round` methods.
 
 Version 1.0.0 (2018-03-03)
 ==========================